+++ laravel-backend/app/Http/Controllers/Admin/ImageController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\ImageRequest;
 use App\Models\Image;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
 use Illuminate\Foundation\Application;
@@ -16,9 +17,13 @@ class ImageController extends Controller
 {
     /**
      * Display a listing of the resource.
+     *
+     * @throws AuthorizationException
      */
     public function index(): View|Application|Factory|\Illuminate\Contracts\Foundation\Application
     {
+        $this->authorize('view images');
+
         return view('images.list', [
             'images' => Image::all(),
         ]);
@@ -26,9 +31,13 @@ class ImageController extends Controller
 
     /**
      * Store a newly created resource in storage.
+     *
+     * @throws AuthorizationException
      */
     public function store(ImageRequest $request): RedirectResponse
     {
+        $this->authorize('add images');
+
         $this->handleSave($request);
 
         return redirect()->back();
@@ -36,17 +45,26 @@ class ImageController extends Controller
 
     /**
      * Remove the specified resource from storage.
+     *
+     * @throws AuthorizationException
      */
     public function destroy(Image $image): RedirectResponse
     {
+        $this->authorize('delete images');
+
         Storage::delete($image->path);
         $image->delete();
 
         return redirect()->back();
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function editorImageUpload(ImageRequest $request): JsonResponse
     {
+        $this->authorize('add images');
+
         $image = $this->handleSave($request);
 
         return response()->json([
